# Conflicts:
#	club_management/admins/forms.py
#	club_management/admins/method.py
#	club_management/admins/signals.py
#	club_management/admins/templates/admins/manage report/report.html
#	club_management/admins/templates/admins/manage request/user request.html
#	club_management/admins/templates/admins/manage task/edit task detail.html
#	club_management/admins/views.py
#	club_management/club/templates/club/base.html
#	club_management/club/templates/club/home.html
#	club_management/club/views.py
#	club_management/db.sqlite3
#	club_management/users/signals.py
#	club_management/users/views.py

+++ club_management/admins/signals.py
@@ -1,11 +1,32 @@
-from django.db.models.signals import pre_delete
+from django.db.models.signals import pre_delete, pre_save
 from django.dispatch import receiver
 from users.models import Task_assigned
+from .models import Page
 from club_management.settings import MEDIA_URL
 import os
 
 
 @receiver(pre_delete, sender=Task_assigned)
-def delete_profile(sender, instance, **kwargs):
-    if os.path.exists(f'{MEDIA_URL.replace("/", "")}/{instance.upload_file}'):
-        os.remove(f'{MEDIA_URL.replace("/", "")}/{instance.upload_file}')
+def delete_task_file(sender, instance, **kwargs):
+    if instance.upload_file:
+        if os.path.exists(f'{MEDIA_URL.replace("/", "")}/{instance.upload_file}'):
+            os.remove(f'{MEDIA_URL.replace("/", "")}/{instance.upload_file}')
+
+
+@receiver(pre_save, sender=Page)
+def replace_image(sender, instance, **kwargs):
+    page = Page.objects.first()
+    if page is None:
+        return
+
+    if not(str(instance.top_background).startswith('page')):
+        page = Page.objects.get(id=instance.id)
+        if page.top_background:
+            if os.path.exists(f'{MEDIA_URL.replace("/", "")}/{page.top_background}'):
+                os.remove(f'{MEDIA_URL.replace("/", "")}/{page.top_background}')
+
+    if not(str(instance.image).startswith('page')):
+        page = Page.objects.get(id=instance.id)
+        if page.image:
+            if os.path.exists(f'{MEDIA_URL.replace("/", "")}/{page.image}'):
+                os.remove(f'{MEDIA_URL.replace("/", "")}/{page.image}')
